updated cheatmenu_din.qsrc: Added lactation on off switch to the status menu as well as some things to change when lactation is on. Tested everything for an hour. seems fine.
updated lact_lib.qsrc
Change in calc_lactatemm as inducing isn't doing anything here.
Added $get_breastmilk_time function. This funktion can be used to calculate how long it would take to get a specific amount of milk from svetas breasts. This is basically a reverse version of the the get breastmilk function.
Made some clean up in other functions and tweaked an exception in if/else function as one case was not handled, but is now.