Updated regex and new lowercase changes.
Command used (thanks to the wizards from freenode's #regex).
find locations/ -type f -name *.qsrc -exec perl -i -pe 's/(?>src="|\G<<(?:[^>"]+|>(?!>))*>>|\G'\''\s*\+(?:[^<"'\''+]+|\+(?!\s*'\''))+(?:\+\s*'\'')?)\K(?>[^<"'\'']+|<(?!<))+/\L$&/g' {} \;

This one might fail on specific use case, if any of the non-string
portions contain < " or '. Otherwise will try to properly apply a
lowercase around the <<>> tags and the string concatenations.

NOTE: The bit of string '\'' is intrinsic to bash. Single quotes can't be
escaped in an intuitive maner. As such, if the regex needs to be used in
another environment, make sure to replace the '\'' group with a single '.