Add in Tree Circle teleport locations and events
This is a preliminary add in.  It is not yet working.  I have added a stub in to each park/forest location ot get to the treeCricle, and a stub for the event to get the teleport spell.  The user should be able to teleport to locations he has found.